Rename file to ComparisonTestDbHelpers.ts
