+++ packages/core/phase2/phase2.comparison.test.ts
@@ -15,8 +15,8 @@ import CoreAuditLogger from "../lib/auditLog/CoreAuditLogger"
 import saveErrorListRecord from "../lib/database/saveErrorListRecord"
 import serialiseToXml from "../lib/serialise/pncUpdateDatasetXml/serialiseToXml"
 import checkDatabaseMatches from "../tests/helpers/comparison/checkDatabaseMatches"
+import { clearDatabase, disconnectDb, insertRecords, sql } from "../tests/helpers/comparison/ComparisonTestDbHelpers"
 import convertXmlAuditLogs from "../tests/helpers/comparison/convertXmlAuditLogs"
-import { clearDatabase, disconnectDb, insertRecords, sql } from "../tests/helpers/comparison/e2eComparisonTestsHelpers"
 import getComparisonTests from "../tests/helpers/comparison/getComparisonTests"
 import parseIncomingMessage from "../tests/helpers/comparison/parseIncomingMessage"
 import sortTriggers from "../tests/helpers/sortTriggers"
