+++ packages/core/tests/helpers/comparison/checkDatabaseMatches.ts
@@ -2,7 +2,7 @@ import type ErrorListRecord from "../../../types/ErrorListRecord"
 import type ErrorListTriggerRecord from "../../../types/ErrorListTriggerRecord"
 import type { DbRecords } from "../../types/ComparisonFile"
 
-import { sql } from "./e2eComparisonTestsHelpers"
+import { sql } from "./ComparisonTestDbHelpers"
 import normaliseErrorListTriggers from "./normaliseErrorListTriggers"
 import normaliseXml from "./normaliseXml"
